Move functions to update widgets / config to a new module 'sync'
